+++ src/QSymbolicsBase/basic_ops_homogeneous.jl
@@ -22,7 +22,6 @@ julia> 2*A
 @withmetadata struct SScaled{T<:QObj} <: Symbolic{T}
     coeff
     obj
-    SScaled{S}(c,k) where S = _isone(c) ? k : new{S}(c,k)
 end
 isexpr(::SScaled) = true
 iscall(::SScaled) = true
@@ -31,10 +30,14 @@ operation(x::SScaled) = *
 head(x::SScaled) = :*
 children(x::SScaled) = [:*,x.coeff,x.obj]
 function Base.:(*)(c, x::Symbolic{T}) where {T<:QObj} 
-    if iszero(c) || iszero(x)
+    if (isa(c, Number) && iszero(c)) || iszero(x)
         SZero{T}()
+    elseif _isone(c)
+        x
+    elseif isa(x, SScaled)
+        SScaled{T}(c*x.coeff, x.obj)
     else 
-        x isa SScaled ? SScaled{T}(c*x.coeff, x.obj) : SScaled{T}(c, x) 
+        SScaled{T}(c, x) 
     end
 end
 Base.:(*)(x::Symbolic{T}, c) where {T<:QObj} = c*x
@@ -81,8 +84,8 @@ julia> k₁ + k₂
     _arguments_precomputed
 end
 function SAdd{S}(d) where S 
-    xs = [c*obj for (c,obj) in d]
-    length(d)==1 ? first(xs) : SAdd{S}(d,Set(xs),xs)
+    terms = [c*obj for (obj,c) in d]
+    length(d)==1 ? first(terms) : SAdd{S}(d,Set(terms),terms)
 end
 isexpr(::SAdd) = true
 iscall(::SAdd) = true
@@ -99,6 +102,11 @@ end
 Base.:(+)(xs::Vararg{Symbolic{<:QObj},0}) = 0 # to avoid undefined type parameters issue in the above method
 basis(x::SAdd) = basis(first(x.dict).first)
 
+const SAddBra = SAdd{AbstractBra}
+function Base.show(io::IO, x::SAddBra)
+    ordered_terms = sort([repr(i) for i in arguments(x)])
+    print(io, "("*join(ordered_terms,"+")::String*")") # type assert to help inference
+end
 const SAddKet = SAdd{AbstractKet}
 function Base.show(io::IO, x::SAddKet)
     ordered_terms = sort([repr(i) for i in arguments(x)])
@@ -106,11 +114,6 @@ function Base.show(io::IO, x::SAddKet)
 end
 const SAddOperator = SAdd{AbstractOperator}
 function Base.show(io::IO, x::SAddOperator) 
-    ordered_terms = sort([repr(i) for i in arguments(x)])
-    print(io, "("*join(ordered_terms,"+")::String*")") # type assert to help inference
-end
-const SAddBra = SAdd{AbstractBra}
-function Base.show(io::IO, x::SAddBra)
     ordered_terms = sort([repr(i) for i in arguments(x)])
     print(io, "("*join(ordered_terms,"+")::String*")") # type assert to help inference
 end
@@ -126,10 +129,6 @@ AB
 """
 @withmetadata struct SMulOperator <: Symbolic{AbstractOperator}
     terms
-    function SMulOperator(terms)
-        coeff, cleanterms = prefactorscalings(terms)
-        coeff*new(cleanterms)
-    end
 end
 isexpr(::SMulOperator) = true
 iscall(::SMulOperator) = true
@@ -139,7 +138,13 @@ head(x::SMulOperator) = :*
 children(x::SMulOperator) = [:*;x.terms]
 function Base.:(*)(xs::Symbolic{AbstractOperator}...) 
     zero_ind = findfirst(x->iszero(x), xs)
-    isnothing(zero_ind) ? SMulOperator(collect(xs)) : SZeroOperator()
+    if isnothing(zero_ind)
+        terms = flattenop(*, collect(xs))
+        coeff, cleanterms = prefactorscalings(terms)
+        coeff * SMulOperator(cleanterms)
+    else
+        SZeroOperator()
+    end
 end
 Base.show(io::IO, x::SMulOperator) = print(io, join(map(string, arguments(x)),""))
 basis(x::SMulOperator) = basis(x.terms)
@@ -155,25 +160,27 @@ julia> k₁ ⊗ k₂
 julia> @op A; @op B;
 
 julia> A ⊗ B 
-A⊗B
+(A⊗B)
 ```
 """
 @withmetadata struct STensor{T<:QObj} <: Symbolic{T}
     terms
-    function STensor{S}(terms) where S
-        coeff, cleanterms = prefactorscalings(terms)
-        coeff * new{S}(cleanterms)
-    end
 end
 isexpr(::STensor) = true
 iscall(::STensor) = true
 arguments(x::STensor) = x.terms
 operation(x::STensor) = ⊗
 head(x::STensor) = :⊗
-children(x::STensor) = pushfirst!(x.terms,:⊗)
+children(x::STensor) = [:⊗; x.terms]
 function ⊗(xs::Symbolic{T}...) where {T<:QObj}
     zero_ind = findfirst(x->iszero(x), xs)
-    isnothing(zero_ind) ? STensor{T}(collect(xs)) : SZero{T}()
+    if isnothing(zero_ind)
+        terms = flattenop(⊗, collect(xs))
+        coeff, cleanterms = prefactorscalings(terms)
+        coeff * STensor{T}(cleanterms)
+    else
+        SZero{T}()
+    end
 end
 basis(x::STensor) = tensor(basis.(x.terms)...)
 
@@ -182,9 +189,9 @@ Base.show(io::IO, x::STensorBra) = print(io, join(map(string, arguments(x)),""))
 const STensorKet = STensor{AbstractKet}
 Base.show(io::IO, x::STensorKet) = print(io, join(map(string, arguments(x)),""))
 const STensorOperator = STensor{AbstractOperator}
-Base.show(io::IO, x::STensorOperator) = print(io, join(map(string, arguments(x)),"⊗"))
+Base.show(io::IO, x::STensorOperator) = print(io, "("*join(map(string, arguments(x)),"⊗")*")")
 const STensorSuperOperator = STensor{AbstractSuperOperator}
-Base.show(io::IO, x::STensorSuperOperator) = print(io, join(map(string, arguments(x)),"⊗"))
+Base.show(io::IO, x::STensorSuperOperator) = print(io, "("*join(map(string, arguments(x)),"⊗")*")")
 
 """Symbolic commutator of two operators
 
@@ -201,10 +208,6 @@ julia> commutator(A, A)
 @withmetadata struct SCommutator <: Symbolic{AbstractOperator}
     op1
     op2
-    function SCommutator(o1, o2) 
-        coeff, cleanterms = prefactorscalings([o1 o2], scalar=true)
-        cleanterms[1] === cleanterms[2] ? SZeroOperator() : coeff*new(cleanterms...)
-    end
 end
 isexpr(::SCommutator) = true
 iscall(::SCommutator) = true
@@ -212,13 +215,15 @@ arguments(x::SCommutator) = [x.op1, x.op2]
 operation(x::SCommutator) = commutator
 head(x::SCommutator) = :commutator
 children(x::SCommutator) = [:commutator, x.op1, x.op2]
-commutator(o1::Symbolic{AbstractOperator}, o2::Symbolic{AbstractOperator}) = SCommutator(o1, o2)
+function commutator(o1::Symbolic{AbstractOperator}, o2::Symbolic{AbstractOperator})
+    coeff, cleanterms = prefactorscalings([o1 o2])
+    cleanterms[1] === cleanterms[2] ? SZeroOperator() : coeff * SCommutator(cleanterms...)   
+end
 commutator(o1::SZeroOperator, o2::Symbolic{AbstractOperator}) = SZeroOperator()
 commutator(o1::Symbolic{AbstractOperator}, o2::SZeroOperator) = SZeroOperator()
 commutator(o1::SZeroOperator, o2::SZeroOperator) = SZeroOperator()
 Base.show(io::IO, x::SCommutator) = print(io, "[$(x.op1),$(x.op2)]")
 basis(x::SCommutator) = basis(x.op1)
-expand(x::SCommutator) = x == 0 ? x : x.op1*x.op2 - x.op2*x.op1
 
 """Symbolic anticommutator of two operators
 
@@ -232,10 +237,6 @@ julia> anticommutator(A, B)
 @withmetadata struct SAnticommutator <: Symbolic{AbstractOperator}
     op1
     op2
-    function SAnticommutator(o1, o2) 
-        coeff, cleanterms = prefactorscalings([o1 o2], scalar=true)
-        coeff*new(cleanterms...)
-    end
 end
 isexpr(::SAnticommutator) = true
 iscall(::SAnticommutator) = true
@@ -243,10 +244,12 @@ arguments(x::SAnticommutator) = [x.op1, x.op2]
 operation(x::SAnticommutator) = anticommutator
 head(x::SAnticommutator) = :anticommutator
 children(x::SAnticommutator) = [:anticommutator, x.op1, x.op2]
-anticommutator(o1::Symbolic{AbstractOperator}, o2::Symbolic{AbstractOperator}) = SAnticommutator(o1, o2)
+function anticommutator(o1::Symbolic{AbstractOperator}, o2::Symbolic{AbstractOperator})
+    coeff, cleanterms = prefactorscalings([o1 o2])
+    coeff * SAnticommutator(cleanterms...)
+end
 anticommutator(o1::SZeroOperator, o2::Symbolic{AbstractOperator}) = SZeroOperator()
 anticommutator(o1::Symbolic{AbstractOperator}, o2::SZeroOperator) = SZeroOperator()
 anticommutator(o1::SZeroOperator, o2::SZeroOperator) = SZeroOperator()
 Base.show(io::IO, x::SAnticommutator) = print(io, "{$(x.op1),$(x.op2)}")
 basis(x::SAnticommutator) = basis(x.op1)
-expand(x::SAnticommutator) = x == 0 ? x : x.op1*x.op2 + x.op2*x.op1