+++ src/QSymbolicsBase/rules.jl
@@ -1,5 +1,5 @@
 ##
-# This file defines automatic simplification rules for specific operations of quantum objects.
+# This file defines manual simplification and expansion rules for specific operations of quantum objects.
 ##
 
 ##
@@ -11,60 +11,11 @@ function hasscalings(xs)
     end
 end
 _isa(T) = x->isa(x,T)
-
-##
-# Determining factors for expressions containing quantum objects
-##
-
-function prefactorscalings(xs; scalar=false) # If the scalar keyword is true, then only scalar factors will be returned as coefficients
-    terms = []
-    coeff = 1::Any
-    for x in xs
-        if isexpr(x) && operation(x) == *
-            c,t = arguments(x)
-            if !scalar
-                coeff *= c
-                push!(terms,t)
-            elseif scalar && c isa Number
-                coeff *= c
-                push!(terms, t)
-            else
-                push!(terms,(*)(arguments(x)...))
-            end
-        else
-            push!(terms,x)
-        end
-    end
-    coeff, terms
-end
-
-function prefactorscalings_rule(xs)
-    coeff, terms = prefactorscalings(xs)
-    coeff * ⊗(terms...)
-end
-
-function isnotflat_precheck(*)
-    function (x)
-        operation(x) === (*) || return false
-        args = arguments(x)
-        for t in args
-            if isexpr(t) && operation(t) === (*)
-                return true
-            end
-        end
-        return false
-    end
-end
+_vecisa(T) = x->all(_isa(T), x)
 
 ##
 # Simplification rules
 ## 
-
-# Flattening expressions
-RULES_FLATTEN = [
-    @rule(~x::isnotflat_precheck(⊗) => flatten_term(⊗, ~x)),
-    @rule ⊗(~~xs::hasscalings) => prefactorscalings_rule(xs)
-]
 
 # Pauli identities
 RULES_PAULI = [
@@ -105,12 +56,12 @@ RULES_ANTICOMMUTATOR = [
     @rule(anticommutator(~o1::_isa(XGate), ~o2::_isa(ZGate)) => 0)
 ]
 
-RULES_ALL = [RULES_PAULI; RULES_COMMUTATOR; RULES_ANTICOMMUTATOR]
+RULES_SIMPLIFY = [RULES_PAULI; RULES_COMMUTATOR; RULES_ANTICOMMUTATOR]
 
 ##
-# Rewriters
+# Simplification rewriters
 ##
-qsimplify_flatten = Chain(RULES_FLATTEN)
+
 qsimplify_anticommutator = Chain(RULES_ANTICOMMUTATOR)
 qsimplify_pauli = Chain(RULES_PAULI)
 qsimplify_commutator = Chain(RULES_COMMUTATOR)
@@ -131,10 +82,57 @@ julia> qsimplify(anticommutator(σˣ, σˣ), rewriter=qsimplify_anticommutator)
 function qsimplify(s; rewriter=nothing)
     if QuantumSymbolics.isexpr(s)
         if isnothing(rewriter)
-            Fixpoint(Prewalk(Chain(RULES_ALL)))(s)
+            Fixpoint(Prewalk(Chain(RULES_SIMPLIFY)))(s)
         else
             Fixpoint(Prewalk(rewriter))(s)
         end
+    else
+        error("Object $(s) of type $(typeof(s)) is not an expression.")
+    end
+end
+
+##
+# Expansion rules
+## 
+
+RULES_EXPAND = [
+    @rule(commutator(~o1, ~o2) => (~o1)*(~o2) - (~o2)*(~o1)),
+    @rule(anticommutator(~o1, ~o2) => (~o1)*(~o2) + (~o2)*(~o1)),
+    @rule(~o1 ⊗ +(~~ops) => +(map(op -> ~o1 ⊗ op, ~~ops)...)),
+    @rule(+(~~ops) ⊗ ~o1 => +(map(op -> op ⊗ ~o1, ~~ops)...)),
+    @rule(~o1 * +(~~ops) => +(map(op -> ~o1 * op, ~~ops)...)),
+    @rule(+(~~ops) * ~o1 => +(map(op -> op * ~o1, ~~ops)...)),
+    @rule(+(~~ops) * ~o1 => +(map(op -> op * ~o1, ~~ops)...)),
+    @rule(⊗(~~ops1::_vecisa(Symbolic{AbstractOperator})) * ⊗(~~ops2::_vecisa(Symbolic{AbstractOperator})) => ⊗(map(*, ~~ops1, ~~ops2)...)),
+    @rule(⊗(~~ops1::_vecisa(Symbolic{AbstractBra})) * ⊗(~~ops2::_vecisa(Symbolic{AbstractKet})) => *(map(*, ~~ops1, ~~ops2)...))
+]
+
+# 
+
+##
+# Expansion rewriter
+##
+
+"""Manually expand a symbolic expression of quantum objects. 
+
+```jldoctest
+julia> @op A; @op B; @op C;
+
+julia> qexpand(commutator(A, B))
+(-1BA+AB)
+
+julia> qexpand(A⊗(B+C))
+((A⊗B)+(A⊗C))
+
+julia> @ket k₁; @ket k₂;
+
+julia> qexpand(A*(k₁+k₂))
+(A|k₁⟩+A|k₂⟩)
+```
+"""
+function qexpand(s)
+    if QuantumSymbolics.isexpr(s)
+        Fixpoint(Prewalk(Chain(RULES_EXPAND)))(s)
     else
         error("Object $(s) of type $(typeof(s)) is not an expression.")
     end